Move diffusion acceleration helpers under optimization
