+++ python/sglang/multimodal_gen/runtime/optimization/__init__.py
@@ -1,0 +1,1 @@
+# SPDX-License-Identifier: Apache-2.0
