+++ python/sglang/multimodal_gen/runtime/server_args.py
@@ -27,7 +27,7 @@ from sglang.multimodal_gen.configs.pipeline_configs.ltx_2 import (
     is_ltx23_native_variant,
 )
 from sglang.multimodal_gen.configs.quantization.nunchaku import NunchakuSVDQuantArgs
-from sglang.multimodal_gen.runtime.acceleration_policy import (
+from sglang.multimodal_gen.runtime.optimization.acceleration_policy import (
     configure_acceleration_policy,
 )
 from sglang.multimodal_gen.runtime.disaggregation.roles import RoleType
